feat(deprecation): removing deprecated Newsfeed, NewsfeedMessage and NewsfeedChangedCallback.
