+++ AndroidSDK/src/com/leanplum/LeanplumInboxMessage.java
@@ -27,11 +27,14 @@ import android.text.TextUtils;
 import com.leanplum.internal.CollectionUtil;
 import com.leanplum.internal.Constants;
 import com.leanplum.internal.Log;
+import com.leanplum.internal.Request;
 import com.leanplum.internal.Util;
 
 import org.json.JSONObject;
 
 import java.io.File;
+import java.util.Date;
+import java.util.HashMap;
 import java.util.Map;
 
 import static com.leanplum.internal.FileManager.DownloadFileResult;
@@ -44,13 +47,22 @@ import static com.leanplum.internal.FileManager.maybeDownloadFile;
  *
  * @author Anna Orlova
  */
-public class LeanplumInboxMessage extends NewsfeedMessage {
+public class LeanplumInboxMessage {
+  private String messageId;
+  private Long deliveryTimestamp;
+  private Long expirationTimestamp;
+  private boolean isRead;
+  private ActionContext context;
   private String imageUrl;
   private String imageFileName;
 
   private LeanplumInboxMessage(String messageId, Long deliveryTimestamp, Long expirationTimestamp,
       boolean isRead, ActionContext context) {
-    super(messageId, deliveryTimestamp, expirationTimestamp, isRead, context);
+    this.messageId = messageId;
+    this.deliveryTimestamp = deliveryTimestamp;
+    this.expirationTimestamp = expirationTimestamp;
+    this.isRead = isRead;
+    this.context = context;
     imageUrl = context.stringNamed(Constants.Keys.INBOX_IMAGE);
     if (imageUrl != null) {
       try {
@@ -58,34 +70,6 @@ public class LeanplumInboxMessage extends NewsfeedMessage {
       } catch (Exception ignored) {
       }
     }
-  }
-
-  static LeanplumInboxMessage createFromJsonMap(String messageId, Map<String, Object> map) {
-    Map<String, Object> messageData = CollectionUtil.uncheckedCast(map.get(Constants.Keys
-        .MESSAGE_DATA));
-    Long deliveryTimestamp = CollectionUtil.uncheckedCast(map.get(Constants.Keys
-        .DELIVERY_TIMESTAMP));
-    Long expirationTimestamp = CollectionUtil.uncheckedCast(map.get(Constants.Keys
-        .EXPIRATION_TIMESTAMP));
-    Boolean isRead = CollectionUtil.uncheckedCast(map.get(Constants.Keys.IS_READ));
-    return constructMessage(messageId, deliveryTimestamp, expirationTimestamp,
-        isRead != null ? isRead : false, messageData);
-  }
-
-  static LeanplumInboxMessage constructMessage(String messageId, Long deliveryTimestamp,
-      Long expirationTimestamp, boolean isRead, Map<String, Object> actionArgs) {
-    if (!isValidMessageId(messageId)) {
-      Log.e("Malformed inbox messageId: " + messageId);
-      return null;
-    }
-
-    String[] messageIdParts = messageId.split("##");
-    ActionContext context = new ActionContext((String) actionArgs.get(Constants.Values.ACTION_ARG),
-        actionArgs, messageIdParts[0]);
-    context.preventRealtimeUpdating();
-    context.update();
-    return new LeanplumInboxMessage(messageId, deliveryTimestamp, expirationTimestamp, isRead,
-        context);
   }
 
   /**
@@ -137,13 +121,131 @@ public class LeanplumInboxMessage extends NewsfeedMessage {
   }
 
   /**
+   * Returns the message identifier of the newsfeed message.
+   */
+  public String getMessageId() {
+    return messageId;
+  }
+
+
+  /**
+   * Returns the title of the inbox message.
+   */
+  public String getTitle() {
+    return context.stringNamed(Constants.Keys.TITLE);
+  }
+
+  /**
+   * Returns the subtitle of the inbox message.
+   */
+  public String getSubtitle() {
+    return context.stringNamed(Constants.Keys.SUBTITLE);
+  }
+
+  /**
+   * Returns the delivery timestamp of the inbox message,
+   * or null if delivery timestamp is not present.
+   */
+  public Date getDeliveryTimestamp() {
+    if (deliveryTimestamp == null) {
+      return null;
+    }
+    return new Date(deliveryTimestamp);
+  }
+
+
+  /**
+   * Return the expiration timestamp of the inbox message.
+   */
+  public Date getExpirationTimestamp() {
+    if (expirationTimestamp == null) {
+      return null;
+    }
+    return new Date(expirationTimestamp);
+  }
+
+  /**
+   * Returns 'true' if the inbox message is read.
+   */
+  public boolean isRead() {
+    return isRead;
+  }
+
+  /**
+   * Read the inbox message, marking it as read and invoking its open action.
+   */
+  public void read() {
+    try {
+      if (Constants.isNoop()) {
+        return;
+      }
+
+      if (!this.isRead) {
+        setIsRead(true);
+
+        int unreadCount = LeanplumInbox.getInstance().unreadCount() - 1;
+        LeanplumInbox.getInstance().updateUnreadCount(unreadCount);
+
+        Map<String, Object> params = new HashMap<>();
+        params.put(Constants.Params.INBOX_MESSAGE_ID, messageId);
+        Request req = Request.post(Constants.Methods.MARK_INBOX_MESSAGE_AS_READ,
+            params);
+        req.send();
+      }
+      this.context.runTrackedActionNamed(Constants.Values.DEFAULT_PUSH_ACTION);
+    } catch (Throwable t) {
+      Util.handleException(t);
+    }
+  }
+
+  /**
+   * Remove the inbox message from the inbox.
+   */
+  public void remove() {
+    try {
+      LeanplumInbox.getInstance().removeMessage(messageId);
+    } catch (Throwable t) {
+      Util.handleException(t);
+    }
+  }
+
+  static LeanplumInboxMessage createFromJsonMap(String messageId, Map<String, Object> map) {
+    Map<String, Object> messageData = CollectionUtil.uncheckedCast(map.get(Constants.Keys
+        .MESSAGE_DATA));
+    Long deliveryTimestamp = CollectionUtil.uncheckedCast(map.get(Constants.Keys
+        .DELIVERY_TIMESTAMP));
+    Long expirationTimestamp = CollectionUtil.uncheckedCast(map.get(Constants.Keys
+        .EXPIRATION_TIMESTAMP));
+    Boolean isRead = CollectionUtil.uncheckedCast(map.get(Constants.Keys.IS_READ));
+    return constructMessage(messageId, deliveryTimestamp, expirationTimestamp,
+        isRead != null ? isRead : false, messageData);
+  }
+
+  static LeanplumInboxMessage constructMessage(String messageId, Long deliveryTimestamp,
+      Long expirationTimestamp, boolean isRead, Map<String, Object> actionArgs) {
+    if (!isValidMessageId(messageId)) {
+      Log.e("Malformed inbox messageId: " + messageId);
+      return null;
+    }
+
+    String[] messageIdParts = messageId.split("##");
+    ActionContext context = new ActionContext((String) actionArgs.get(Constants.Values.ACTION_ARG),
+        actionArgs, messageIdParts[0]);
+    context.preventRealtimeUpdating();
+    context.update();
+    return new LeanplumInboxMessage(messageId, deliveryTimestamp, expirationTimestamp, isRead,
+        context);
+  }
+
+
+  /**
    * Download image if prefetching is enabled.
    * Uses {@link LeanplumInbox#downloadedImageUrls} to make sure we don't call fileExist method
    * multiple times for same URLs.
    *
    * @return Boolean True if the image will be downloaded, otherwise false.
    */
-  Boolean downloadImageIfPrefetchingEnabled() {
+  boolean downloadImageIfPrefetchingEnabled() {
     if (!LeanplumInbox.isInboxImagePrefetchingEnabled) {
       return false;
     }
@@ -157,4 +259,39 @@ public class LeanplumInboxMessage extends NewsfeedMessage {
     LeanplumInbox.downloadedImageUrls.add(imageUrl);
     return DownloadFileResult.DOWNLOADING == result;
   }
+
+  Map<String, Object> toJsonMap() {
+    Map<String, Object> map = new HashMap<>();
+    map.put(Constants.Keys.DELIVERY_TIMESTAMP, this.deliveryTimestamp);
+    map.put(Constants.Keys.EXPIRATION_TIMESTAMP, this.expirationTimestamp);
+    map.put(Constants.Keys.MESSAGE_DATA, this.actionArgs());
+    map.put(Constants.Keys.IS_READ, this.isRead());
+    return map;
+  }
+
+  boolean isActive() {
+    if (expirationTimestamp == null) {
+      return true;
+    }
+
+    Date now = new Date();
+    return now.before(new Date(expirationTimestamp));
+  }
+
+  private static boolean isValidMessageId(String messageId) {
+    return messageId.split("##").length == 2;
+  }
+
+  ActionContext getContext() {
+    return context;
+  }
+
+  private Map<String, Object> actionArgs() {
+    return context.getArgs();
+  }
+
+  private void setIsRead(boolean isRead) {
+    this.isRead = isRead;
+  }
+
 }
